requirements.txt
openai
pyaudio
websockets
python-dotenv
whisper
